Having head => None gives None when resolving functions
This is an experimental hack to understand how interpretation changes if we
change the algorithm for resolving functions.

The original implementation resolved the Values first, discarding any Nones and
then applied the function.

This change assumes that if the first value in the args list (that determines
the "type" of the application) is None, then we should yield Nones. This allows
for calculations that should "do nothing if something is missing". Thus, this
avoids implementing branching concepts of the form:

MAP a=or(fn0(...), fnN(...)) where the or seeks the FIRST fnN that yields a
value.